Move tests  for isna() into a separate file
