+++ docs/api/math/isna.rst
@@ -2,3 +2,4 @@
 .. xfunction:: datatable.math.isna
     :src: src/core/expr/funary/floating.cc resolve_op_isna
     :doc: src/core/expr/funary/floating.cc doc_isna
+    :tests: tests/expr/test-isna.py
