+++ tests/ijby/test-math-floating.py
@@ -348,47 +348,6 @@ def test_isinf_srcs(src):
 
 
 #-------------------------------------------------------------------------------
-# math.isna()
-#-------------------------------------------------------------------------------
-
-@pytest.mark.parametrize("src", srcs_bool + srcs_int + srcs_float + srcs_str)
-def test_dt_isna(src):
-    DT = dt.Frame(src)
-    RES = DT[:, dt.math.isna(f[0])]
-    assert_equals(RES, dt.Frame([(x is None) for x in src]))
-
-
-def test_dt_isna2():
-    from math import nan
-    DT = dt.Frame(A=[1, None, 2, 5, None, 3.6, nan, -4.899])
-    DT1 = DT[~dt.math.isna(f.A), :]
-    assert DT1.stypes == DT.stypes
-    assert DT1.names == DT.names
-    assert DT1.to_list() == [[1.0, 2.0, 5.0, 3.6, -4.899]]
-
-
-def test_dt_isna_joined():
-    # See issue #2109
-    DT = dt.Frame(A=[None, 4, 3, 2, 1])
-    JDT = dt.Frame(A=[0, 1, 3, 7],
-                   B=['a', 'b', 'c', 'd'],
-                   C=[0.25, 0.5, 0.75, 1.0],
-                   D=[22, 33, 44, 55],
-                   E=[True, False, True, False])
-    JDT.key = 'A'
-    RES = DT[:, dt.math.isna(g[1:]), join(JDT)]
-    frame_integrity_check(RES)
-    assert RES.to_list() == [[True, True, False, True, False]] * 4
-
-
-@pytest.mark.parametrize("src", srcs_bool + srcs_int + srcs_float + srcs_str)
-def test_dt_math_isna_scalar(src):
-    for val in src:
-        assert dt.math.isna(val) == (val is None or val is math.nan)
-
-
-
-#-------------------------------------------------------------------------------
 # math.ldexp()
 #-------------------------------------------------------------------------------
